e2e tests: Dump namespace content on failure (#939)
Print all JobSets and Jobs in the testing namespace after every test
run. This is only visible on failure and it's disabled by default.

A disadvantage is that the output is generated in AfterEach whether the
test fails or not. This only happens 9 times currently and it doesn't
seem to be much of an overhead.

The output is shown only for the failing test, though.

To enable this extended output, set JOBSET_E2E_TEST_DUMP_NAMESPACE
environment variable to "true". This is set by default in
hacks/e2e-tests.sh so that it's visible on CI.