jobset: Don't create jobs/service when suspended
Currently when a JobSet is created suspended, the associated jobs and
headless service are always created immediately, although the jobs too
are suspended.

This patch changes the behavior to only create these child resources
when the JobSet is unsuspended for the first time, hence saving
resources then the JobSet spends a long time suspended before being
started, which can happen with Kueue.